requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
gpxpy==1.5.0
idna==3.4
python-srtm==0.5.0
requests==2.28.2
SRTM.py==0.3.7
urllib3==1.26.14
